extract test fixture to separate module
